Tdd configuration and a small fix for get_band() for FR2

+++ openair2/LAYER2/NR_MAC_UE/config_ue.c
@@ -205,7 +205,7 @@ void config_common_ue(NR_UE_MAC_INST_t *mac,
   uint16_t band;
   int32_t offset;
 
-  get_band((cfg->carrier_config.dl_frequency)*1000,
+  get_band((uint64_t)(cfg->carrier_config.dl_frequency)*1000,
            &band,
            &offset,
            &frame_type);
